+++ crystal/__init__.py
@@ -1,5 +1,5 @@
 __version__ = "1.0.0"
-__build__ = "018"
-__release__ = "Crystal V1.00 Build 018"
+__build__ = "019"
+__release__ = "Crystal V1.00 Build 019"
 
 __all__ = ["__version__", "__build__", "__release__"]